canvas_editor/requirements.txt
django
djangorestframework
git+https://github.com/ARTIST-Association/ARTIST.git
torch
django-cleanup
django-allauth
requests
PyJWT
python-dotenv
cryptography
Pillow
coverage
mypy
pre-commit
ruff
parameterized
unittest-xml-reporting
